requirements.txt
psycopg2-binary
pandas
matplotlib
seaborn
sqlalchemy